fix: code adjustments based on comments
