+++ src/services/attribute/types.ts
@@ -1,0 +1,9 @@
+export interface IGroupedAttribute {
+  name: string;
+  options: string[];
+}
+
+export interface IGroupedAttributesResult {
+  attributeNames: string[];
+  attributeOptions: Record<string, string[]>;
+}
